TEST: Reorganise test logic
Tests that focus on side-effects (transitionary state) will now also account
for the existence of errors when executing state functions (i.e. they account
for both output and effect). The test for activeJobs has been removed since
there was too much overlap between the implementation of the test and the
function itself. Tests and the utility functions on which they depend have
also been modified to make the intent of each test more clear.